report login and data

+++ app/src/main/java/zzbcar/cckj/com/nzzb/bean/CarDetailBean.java
@@ -64,7 +64,7 @@ public class CarDetailBean implements Serializable {
         private String pics;
         private String plateNo;
         private double price;
-        private int receivePercent;
+        private double receivePercent;
         private int seatNum;
         private int series;
         private String seriesName;
@@ -265,11 +265,11 @@ public class CarDetailBean implements Serializable {
             this.price = price;
         }
 
-        public int getReceivePercent() {
+        public double getReceivePercent() {
             return receivePercent;
         }
 
-        public void setReceivePercent(int receivePercent) {
+        public void setReceivePercent(double receivePercent) {
             this.receivePercent = receivePercent;
         }
